FIX: Change to path params instead of query params in APIs
For the download and upload API we are switching to path params
instead of query params. Query params are still used for the query
route to search the database.
