+++ tests/lambda_endpoints/test_download_api.py
@@ -2,14 +2,14 @@ from pathlib import Path
 
 import pytest
 
-from sds_data_manager.lambda_code.SDSCode.download_query_api import lambda_handler
+from sds_data_manager.lambda_code.SDSCode.download_api import lambda_handler
 
 BUCKET_NAME = "test-bucket"
 TEST_FILE = "science_block_20221116_163611Z_idle.bin"
 
 
 @pytest.fixture(autouse=True)
-def setup_s3(s3_client):
+def setup_s3(s3_client, monkeypatch):
     """Populate the mocked s3 client with a bucket and a file
 
     Each test below will use this fixture by default
@@ -20,6 +20,8 @@ def setup_s3(s3_client):
     result = s3_client.list_buckets()
     assert len(result["Buckets"]) == 1
     assert result["Buckets"][0]["Name"] == BUCKET_NAME
+    # Temporarily set the environment variable S3_BUCKET for the test
+    monkeypatch.setenv("S3_BUCKET", BUCKET_NAME)
 
     # upload a file
     local_filepath = Path(__file__).parent.parent.resolve() / f"test-data/{TEST_FILE}"
@@ -35,10 +37,8 @@ def test_object_exists_with_s3_uri():
         "version": "2.0",
         "routeKey": "$default",
         "rawPath": "/",
-        "rawQueryString": f"s3_uri=s3://{BUCKET_NAME}/{TEST_FILE}",
-        "queryStringParameters": {"s3_uri": f"s3://{BUCKET_NAME}/{TEST_FILE}"},
+        "pathParameters": {"proxy": TEST_FILE},
     }
-
     response = lambda_handler(event=event, context=None)
     assert response["statusCode"] == 302
     assert "Location" in response["headers"]
@@ -51,10 +51,7 @@ def test_object_exists_with_s3_uri_fails():
         "version": "2.0",
         "routeKey": "$default",
         "rawPath": "/",
-        "rawQueryString": f"s3_uri=s3://{BUCKET_NAME}/bad_path/bad_file.txt",
-        "queryStringParameters": {
-            "s3_uri": f"s3://{BUCKET_NAME}/bad_path/bad_file.txt"
-        },
+        "pathParameters": {"proxy": "bad_path/bad_file.txt"},
     }
 
     response = lambda_handler(event=event, context=None)
@@ -67,19 +64,8 @@ def test_input_parameters_missing():
         "version": "2.0",
         "routeKey": "$default",
         "rawPath": "/",
-        "rawQueryString": "",
-    }
-
-    bad_para_event = {
-        "version": "2.0",
-        "routeKey": "$default",
-        "rawPath": "/",
-        "rawQueryString": f"bad_input={TEST_FILE}",
-        "queryStringParameters": {"bad_input": f"{TEST_FILE}"},
+        # No pathParameters
     }
 
     response = lambda_handler(event=empty_para_event, context=None)
-    assert response["statusCode"] == 400
-
-    response = lambda_handler(event=bad_para_event, context=None)
     assert response["statusCode"] == 400
